requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
httpx>=0.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pyhdfs>=0.3.1
faiss-cpu>=1.7.4
spacy>=3.6.0
fr_core_news_sm>=3.6.0
numpy>=1.24.0
pydantic>=2.0.0 